+++ Events/src/main/java/org/example/dtos/sessionTiers/SessionTierDTO.java
@@ -1,7 +1,6 @@
 package org.example.dtos.sessionTiers;
 
 
-import java.math.BigDecimal;
 import java.util.Date;
 import java.util.UUID;
 
@@ -13,7 +12,7 @@ public class SessionTierDTO {
 
     private UUID zoneId;
 
-    private BigDecimal price;
+    private float price;
 
     private UUID createdBy;
